requirements.txt
streamlit
google-generativeai
python-dotenv
openpyxl
langchain
cryptography
faker